Reproducible builds (#8)
* Implemented reproducible builds

* Updated build scripts

* Documented AV false positives

* Update README.md

* Update release.yml

* Build action drafts checksums

* Update release.yml

* build.ps1 output utf8

Authored-by: Nathan Beals <ndbeals@users.noreply.github.com>

+++ build.ps1
@@ -37,7 +37,13 @@ if ($Release)
     Copy-Item LICENSE $outDir
     
     Remove-Item -LiteralPath $ReleasePath -ErrorAction SilentlyContinue
+
+    Write-Output "
+## Checksums
+| Architecture | Checksum |
+|---|---|" | Out-File -FilePath checksums.md -Encoding utf8 
 }
+
 # Build for each architecture
 Foreach ($arch in $Architectures)
 {
@@ -45,23 +51,28 @@ Foreach ($arch in $Architectures)
     
     if ($Release)
     {        
-        go build -ldflags -H=windowsgui -o $outDir\winssh-pageant.exe
+        go build -ldflags -H=windowsgui -trimpath -o $outDir\winssh-pageant.exe
         if ($LastExitCode -ne 0) { $returnValue = $LastExitCode }
-        # Remove-Item -LiteralPath $ReleasePath -ErrorAction SilentlyContinue
         Compress-Archive -Path $outDir\* -DestinationPath $releaseDir\$ReleasePath-${ver}_$arch.zip -Force
+
+        $hash = (Get-FileHash $outDir\winssh-pageant.exe).Hash 
+        Write-Output "| $arch | $hash |" | Out-File -FilePath checksums.md -Encoding utf8 -Append
         
         Remove-Item -LiteralPath $outDir\winssh-pageant.exe
     } else {
-        go build -ldflags -H=windowsgui -o $outDir\winssh-pageant-$arch.exe
+        go build -ldflags -H=windowsgui -trimpath -o $outDir\winssh-pageant-$arch.exe
     }
 }
-
 
 # Restore env vars
 $env:GOOS = $oldGOOS
 $env:GOARCH = $oldGOARCH
 
 # Cleanup
-Remove-Item -LiteralPath $BuildPath -Force -Recurse -ErrorAction SilentlyContinue
+if ($Release)
+{
+    Write-Output "" | Out-File -FilePath checksums.md -Encoding utf8 -Append
+    Remove-Item -LiteralPath $BuildPath -Force -Recurse -ErrorAction SilentlyContinue
+}
 
 exit $returnValue
